refactor(store): split tree-replace diff predicates into RecordingStore.TreeReplaceDiff.cs
